\name{Griffith}
\docType{data}
\alias{Griffith}

\title{Example Data from Mulaik (2018)}

\description{

  Example from Mulaik, S.A. (2018) Fundamentals of Common Analysis in 
  Irwing, P., Booth, T., Hughes, D. J.  (2018). The Wiley Handbook of 
  Psychometric Testing, 2 Volume Set: A Multidisciplinary Reference on Survey, 
  Scale and Test Development. [VitalSource Bookshelf 10.2.0]. 
  Retrieved from vbk://9781118489703
  See Part II, Table 8.5 (p.225) and Table 8.6 (p.235)}
\usage{
	data(Griffith)
}
\details{
	The object Griffith is loaded from the data file Griffith.
}

\format{
	The object Griffith is a matrix.
}

\source{
  Mulaik, S.A. (2018) Fundamentals of Common Analysis in 
  Irwing, P., Booth, T., Hughes, D. J.  (2018). The Wiley Handbook of 
  Psychometric Testing
}

\seealso{
	\code{\link{ff.varimax}}
}
\keyword{datasets}
